[UPDATE] 타입 수정

+++ src/main/java/com/dsmpear/main/user_backend_v2/entity/report/ReportRepository.java
@@ -1,17 +1,20 @@
 package com.dsmpear.main.user_backend_v2.entity.report;
 
+import com.dsmpear.main.user_backend_v2.entity.member.Member;
+import com.dsmpear.main.user_backend_v2.entity.user.User;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
+import java.util.List;
 import java.util.Optional;
 
 
 @Repository
-public interface ReportRepository extends CrudRepository<Report,Integer> {
+public interface ReportRepository extends CrudRepository<Report,Long> {
     // 보고서 갖고오기
-    Optional<Report> findById(Integer reportId);
+    Optional<Report> findById(Long reportId);
 
     //제목 검색 ORM
 //    Page<Report> findAllByReportTypeAccessAndIsAcceptedTrueAndIsSubmittedTrueAndTitleContainsOrderByCreatedAtDesc(Access access, String title, Pageable page);
